refactor errors, add index.js for errors (58)
