+++ frontend/server/errors/index.js
@@ -1,0 +1,4 @@
+import BadRequestError from "./bad-request.js";
+import NotFoundError from "./not-found.js";
+
+export { BadRequestError, NotFoundError }
