+++ frontend/server/errors/not-found.js
@@ -1,0 +1,11 @@
+import { StatusCodes } from 'http-status-codes'
+import CustomAPIError from './custom-api'
+
+class NotFoundError extends CustomAPIError {
+    constructor(message) {
+        super(message)
+        this.statusCode = StatusCodes.NOT_FOUND
+    }
+}
+
+export default NotFoundError
